lmdbs/requirements.txt
numpy>=1.21.2
lmdb>=1.2.1
opencv_python>=4.2.0.32
matplotlib>=3.4.3
Pillow>=8.3.2
scikit_learn>=0.8
